+++ flink-runtime/src/main/java/org/apache/flink/runtime/rpc/akka/AkkaRpcService.java
@@ -18,39 +18,35 @@
 
 package org.apache.flink.runtime.rpc.akka;
 
-import akka.actor.ActorIdentity;
-import akka.actor.ActorRef;
-import akka.actor.ActorSelection;
-import akka.actor.ActorSystem;
-import akka.actor.Address;
-import akka.actor.Cancellable;
-import akka.actor.Identify;
-import akka.actor.Props;
-import akka.dispatch.Futures;
-import akka.dispatch.Mapper;
-import akka.pattern.Patterns;
 import org.apache.flink.api.common.time.Time;
 import org.apache.flink.api.java.tuple.Tuple2;
 import org.apache.flink.runtime.akka.AkkaUtils;
 import org.apache.flink.runtime.concurrent.FutureUtils;
 import org.apache.flink.runtime.concurrent.ScheduledExecutor;
+import org.apache.flink.runtime.concurrent.akka.ActorSystemScheduledExecutorAdapter;
 import org.apache.flink.runtime.rpc.FencedMainThreadExecutable;
 import org.apache.flink.runtime.rpc.FencedRpcEndpoint;
 import org.apache.flink.runtime.rpc.FencedRpcGateway;
 import org.apache.flink.runtime.rpc.RpcEndpoint;
 import org.apache.flink.runtime.rpc.RpcGateway;
-import org.apache.flink.runtime.rpc.RpcService;
 import org.apache.flink.runtime.rpc.RpcServer;
+import org.apache.flink.runtime.rpc.RpcService;
 import org.apache.flink.runtime.rpc.RpcUtils;
 import org.apache.flink.runtime.rpc.exceptions.RpcConnectionException;
-import org.apache.flink.util.Preconditions;
+
+import akka.actor.ActorIdentity;
+import akka.actor.ActorRef;
+import akka.actor.ActorSelection;
+import akka.actor.ActorSystem;
+import akka.actor.Address;
+import akka.actor.Identify;
+import akka.actor.Props;
+import akka.dispatch.Futures;
+import akka.dispatch.Mapper;
+import akka.pattern.Patterns;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import scala.Option;
-import scala.concurrent.Future;
-import scala.concurrent.duration.FiniteDuration;
 
-import javax.annotation.Nonnull;
 import javax.annotation.concurrent.ThreadSafe;
 
 import java.io.Serializable;
@@ -60,13 +56,13 @@ import java.util.HashSet;
 import java.util.Set;
 import java.util.concurrent.Callable;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.Delayed;
 import java.util.concurrent.Executor;
-import java.util.concurrent.FutureTask;
-import java.util.concurrent.RunnableScheduledFuture;
 import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Function;
+
+import scala.Option;
+import scala.concurrent.Future;
 
 import static org.apache.flink.util.Preconditions.checkArgument;
 import static org.apache.flink.util.Preconditions.checkNotNull;
@@ -122,7 +118,7 @@ public class AkkaRpcService implements RpcService {
 			port = -1;
 		}
 
-		internalScheduledExecutor = new InternalScheduledExecutorImpl(actorSystem);
+		internalScheduledExecutor = new ActorSystemScheduledExecutorAdapter(actorSystem);
 	}
 
 	@Override
@@ -419,167 +415,5 @@ public class AkkaRpcService implements RpcService {
 		}, actorSystem.dispatcher());
 
 		return FutureUtils.toJava(resultFuture);
-	}
-
-	/**
-	 * Helper class to expose the internal scheduling logic via a {@link ScheduledExecutor}.
-	 */
-	private static final class InternalScheduledExecutorImpl implements ScheduledExecutor {
-
-		private final ActorSystem actorSystem;
-
-		private InternalScheduledExecutorImpl(ActorSystem actorSystem) {
-			this.actorSystem = Preconditions.checkNotNull(actorSystem, "rpcService");
-		}
-
-		@Override
-		@Nonnull
-		public ScheduledFuture<?> schedule(@Nonnull Runnable command, long delay, @Nonnull TimeUnit unit) {
-			ScheduledFutureTask<Void> scheduledFutureTask = new ScheduledFutureTask<>(command, unit.toNanos(delay), 0L);
-
-			Cancellable cancellable = internalSchedule(scheduledFutureTask, delay, unit);
-
-			scheduledFutureTask.setCancellable(cancellable);
-
-			return scheduledFutureTask;
-		}
-
-		@Override
-		@Nonnull
-		public <V> ScheduledFuture<V> schedule(@Nonnull Callable<V> callable, long delay, @Nonnull TimeUnit unit) {
-			ScheduledFutureTask<V> scheduledFutureTask = new ScheduledFutureTask<>(callable, unit.toNanos(delay), 0L);
-
-			Cancellable cancellable = internalSchedule(scheduledFutureTask, delay, unit);
-
-			scheduledFutureTask.setCancellable(cancellable);
-
-			return scheduledFutureTask;
-		}
-
-		@Override
-		@Nonnull
-		public ScheduledFuture<?> scheduleAtFixedRate(@Nonnull Runnable command, long initialDelay, long period, @Nonnull TimeUnit unit) {
-			ScheduledFutureTask<Void> scheduledFutureTask = new ScheduledFutureTask<>(
-				command,
-				triggerTime(unit.toNanos(initialDelay)),
-				unit.toNanos(period));
-
-			Cancellable cancellable = actorSystem.scheduler().schedule(
-				new FiniteDuration(initialDelay, unit),
-				new FiniteDuration(period, unit),
-				scheduledFutureTask,
-				actorSystem.dispatcher());
-
-			scheduledFutureTask.setCancellable(cancellable);
-
-			return scheduledFutureTask;
-		}
-
-		@Override
-		@Nonnull
-		public ScheduledFuture<?> scheduleWithFixedDelay(@Nonnull Runnable command, long initialDelay, long delay, @Nonnull TimeUnit unit) {
-			ScheduledFutureTask<Void> scheduledFutureTask = new ScheduledFutureTask<>(
-				command,
-				triggerTime(unit.toNanos(initialDelay)),
-				unit.toNanos(-delay));
-
-			Cancellable cancellable = internalSchedule(scheduledFutureTask, initialDelay, unit);
-
-			scheduledFutureTask.setCancellable(cancellable);
-
-			return scheduledFutureTask;
-		}
-
-		@Override
-		public void execute(@Nonnull Runnable command) {
-			actorSystem.dispatcher().execute(command);
-		}
-
-		private Cancellable internalSchedule(Runnable runnable, long delay, TimeUnit unit) {
-			return actorSystem.scheduler().scheduleOnce(
-				new FiniteDuration(delay, unit),
-				runnable,
-				actorSystem.dispatcher());
-		}
-
-		private long now() {
-			return System.nanoTime();
-		}
-
-		private long triggerTime(long delay) {
-			return now() + delay;
-		}
-
-		private final class ScheduledFutureTask<V> extends FutureTask<V> implements RunnableScheduledFuture<V> {
-
-			private long time;
-
-			private final long period;
-
-			private volatile Cancellable cancellable;
-
-			ScheduledFutureTask(Callable<V> callable, long time, long period) {
-				super(callable);
-				this.time = time;
-				this.period = period;
-			}
-
-			ScheduledFutureTask(Runnable runnable, long time, long period) {
-				super(runnable, null);
-				this.time = time;
-				this.period = period;
-			}
-
-			public void setCancellable(Cancellable newCancellable) {
-				this.cancellable = newCancellable;
-			}
-
-			@Override
-			public void run() {
-				if (!isPeriodic()) {
-					super.run();
-				} else if (runAndReset()){
-					if (period > 0L) {
-						time += period;
-					} else {
-						cancellable = internalSchedule(this, -period, TimeUnit.NANOSECONDS);
-
-						// check whether we have been cancelled concurrently
-						if (isCancelled()) {
-							cancellable.cancel();
-						} else {
-							time = triggerTime(-period);
-						}
-					}
-				}
-			}
-
-			@Override
-			public boolean cancel(boolean mayInterruptIfRunning) {
-				boolean result = super.cancel(mayInterruptIfRunning);
-
-				return result && cancellable.cancel();
-			}
-
-			@Override
-			public long getDelay(@Nonnull  TimeUnit unit) {
-				return unit.convert(time - now(), TimeUnit.NANOSECONDS);
-			}
-
-			@Override
-			public int compareTo(@Nonnull Delayed o) {
-				if (o == this) {
-					return 0;
-				}
-
-				long diff = getDelay(TimeUnit.NANOSECONDS) - o.getDelay(TimeUnit.NANOSECONDS);
-				return (diff < 0L) ? -1 : (diff > 0L) ? 1 : 0;
-			}
-
-			@Override
-			public boolean isPeriodic() {
-				return period != 0L;
-			}
-		}
 	}
 }
