+++ flink-runtime/src/main/java/org/apache/flink/runtime/webmonitor/retriever/impl/AkkaJobManagerRetriever.java
@@ -23,6 +23,8 @@ import org.apache.flink.api.java.tuple.Tuple2;
 import org.apache.flink.runtime.akka.AkkaJobManagerGateway;
 import org.apache.flink.runtime.akka.AkkaUtils;
 import org.apache.flink.runtime.concurrent.FutureUtils;
+import org.apache.flink.runtime.concurrent.ScheduledExecutor;
+import org.apache.flink.runtime.concurrent.akka.ActorSystemScheduledExecutorAdapter;
 import org.apache.flink.runtime.instance.ActorGateway;
 import org.apache.flink.runtime.instance.AkkaActorGateway;
 import org.apache.flink.runtime.jobmaster.JobManagerGateway;
@@ -42,31 +44,51 @@ public class AkkaJobManagerRetriever extends LeaderGatewayRetriever<JobManagerGa
 
 	private final ActorSystem actorSystem;
 	private final Time timeout;
+	private final int retries;
+	private final Time retryDelay;
+
+	private final ScheduledExecutor scheduledExecutor;
 
 	public AkkaJobManagerRetriever(
 			ActorSystem actorSystem,
-			Time timeout) {
+			Time timeout,
+			int retries,
+			Time retryDelay) {
 
 		this.actorSystem = Preconditions.checkNotNull(actorSystem);
 		this.timeout = Preconditions.checkNotNull(timeout);
+
+		Preconditions.checkArgument(retries >= 0, "The number of retries must be >= 0.");
+		this.retries = retries;
+
+		this.retryDelay = Preconditions.checkNotNull(retryDelay);
+
+		this.scheduledExecutor = new ActorSystemScheduledExecutorAdapter(actorSystem);
 	}
 
 	@Override
 	protected CompletableFuture<JobManagerGateway> createGateway(CompletableFuture<Tuple2<String, UUID>> leaderFuture) {
-		return leaderFuture.thenCompose(
-			(Tuple2<String, UUID> addressLeaderId) ->
-				FutureUtils.toJava(
-					AkkaUtils.getActorRefFuture(
-						addressLeaderId.f0,
-						actorSystem,
-						FutureUtils.toFiniteDuration(timeout)))
-					.thenApplyAsync(
-						(ActorRef jobManagerRef) -> {
-							ActorGateway leaderGateway = new AkkaActorGateway(
-								jobManagerRef, addressLeaderId.f1);
 
-							return new AkkaJobManagerGateway(leaderGateway);
-						}
-					));
+		return FutureUtils.retryWithDelay(
+			() ->
+				leaderFuture.thenCompose(
+					(Tuple2<String, UUID> addressLeaderId) ->
+						FutureUtils.toJava(
+							AkkaUtils.getActorRefFuture(
+								addressLeaderId.f0,
+								actorSystem,
+								FutureUtils.toFiniteDuration(timeout)))
+							.thenApply(
+								(ActorRef jobManagerRef) -> {
+									ActorGateway leaderGateway = new AkkaActorGateway(
+										jobManagerRef,
+										addressLeaderId.f1);
+
+									return new AkkaJobManagerGateway(leaderGateway);
+								}
+							)),
+			retries,
+			retryDelay,
+			scheduledExecutor);
 	}
 }
